reverse a change in solver

+++ src/Appl/Field.f90
@@ -362,8 +362,8 @@
         double precision :: aa,bb,cc,dd,ee,ff,ss
         double complex :: gg
         double precision, dimension(2) :: xx,yy,zz
-        real*16, dimension(3) :: vv
-        real*16 :: rr
+        real*8, dimension(3) :: vv
+        real*8 :: rr
         real*8 :: gg2
         integer :: n,i0,j0,k0
         double precision :: recfourpi
